project complete

+++ project/app/Http/Controllers/PostController.php
@@ -9,6 +9,15 @@ use Illuminate\Support\Str;
 
 class PostController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth')->except([
+            'index',
+            'show'
+            ]);
+        
+    }
+
     public function index()
     {
         return view('posts.index', [
@@ -18,7 +27,8 @@ class PostController extends Controller
 
     public function show(Post $post)
     {
-        return view('posts.show', compact('post'));
+        $posts = Post::where('category_id', $post->category_id)->latest()->limit(6)->get();
+        return view('posts.show', compact('post', 'posts'));
     }
 
     public function create()
@@ -54,16 +64,26 @@ class PostController extends Controller
             // return back();
 
         // * third method to store * //
-            
+
+            $request->validate([
+                'thumbnail' => 'image|mimes:jpeg,png,jpg,svg|max:2048'
+            ]);
+
             /*validate the field *using function method* */ 
             $attr = $request->all();
-
             /* asssign title to the slug */
-            $attr['slug'] = Str::slug(request('title'));
+            $slug = Str::slug(request('title'));
+            $attr['slug'] = $slug;
+
+            $thumbnail = request()->file('thumbnail') ? request()->file('thumbnail')->store("images/posts") : null;
+            
+            
+            /* assign data to field */
             $attr['category_id'] = request('category');
+            $attr['thumbnail'] = $thumbnail;
 
             /* create new post */
-            $post = Post::create($attr);
+            $post = auth()->user()->posts()->create($attr);
 
             $post->tags()->attach(request('tags'));
 
@@ -90,11 +110,28 @@ class PostController extends Controller
 
     public function update( PostRequest $request, Post $post)
     {
+        $request->validate([
+            'thumbnail' => 'image|mimes:jpeg,png,jpg,svg|max:2048'
+        ]);
+
+        $this->authorize('update', $post);
+
+        if(request()->file('thumbnail')){
+            \Storage::delete($post->thumbnail);
+            $thumbnail = request()->file('thumbnail')->store("images/posts");
+        } else {
+            $thumbnail = $post->thumbnail;
+        }
+
+        
+
         /*validate the field  *using PostRequest method* */ 
         $attr = $request->all();
 
         /* update new post */
         $attr['category_id'] = request('category');
+        $attr['thumbnail'] = $thumbnail;
+
         $post->update($attr);
         $post->tags()->sync(request('tags'));
 
@@ -118,13 +155,16 @@ class PostController extends Controller
 
     public function destroy(Post $post)
     {
-        
+        $this->authorize('delete', $post);
+
+        \Storage::delete($post->thumbnail);
         $post->tags()->detach();
         $post->delete();
         
-        session()->flash('success', 'The post was deleted');
-
+        session()->flash('error', "The post was deleted");
         return redirect('posts');
-
+        
+        
+        
     }
 }
